+++ app/src/main/java/org/projectbuendia/client/sync/DbSyncHelper.java
@@ -57,239 +57,12 @@ import java.util.concurrent.ExecutionException;
 /**
  * A helper class for turning the Java beans that are the result of chart RPC calls into
  * appropriate {@link ContentProviderOperation}s for inserting into the DB.
+ *
+ * @deprecated We're moving code to use the {@link org.projectbuendia.client.sync.controllers
+ * .SyncController} interface instead, do not add new code here.
  */
+@Deprecated
 public class DbSyncHelper {
-    private static final Logger LOG = Logger.create();
-
-    /** Converts a JsonChart response into appropriate inserts in the chart table. */
-    public static ArrayList<ContentProviderOperation> getChartUpdateOps(
-        JsonChart response, SyncResult syncResult) {
-        if (response.uuid == null) {
-            LOG.e("null chart uuid when fetching chart structure");
-        }
-
-        ArrayList<ContentProviderOperation> ops = new ArrayList<>();
-        int nextId = 1;
-        int nextWeight = 1;
-        for (JsonChartSection section : response.sections) {
-            int parentId = nextId;
-            ops.add(ContentProviderOperation.newInsert(ChartItems.CONTENT_URI)
-                .withValue("rowid", nextId++)
-                .withValue(ChartItems.CHART_UUID, response.uuid)
-                .withValue(ChartItems.WEIGHT, nextWeight++)
-                .withValue(ChartItems.SECTION_TYPE, section.type == null ? null : section.type.name())
-                .withValue(ChartItems.LABEL, section.label)
-                .build());
-            syncResult.stats.numInserts++;
-
-            for (JsonChartItem item : section.items) {
-                Object[] conceptUuids = new Object[item.concepts.length];
-                for (int i = 0; i < conceptUuids.length; i++) {
-                    conceptUuids[i] = Utils.expandUuid(item.concepts[i]);
-                }
-                ops.add(ContentProviderOperation.newInsert(ChartItems.CONTENT_URI)
-                    .withValue("rowid", nextId++)
-                    .withValue(ChartItems.CHART_UUID, response.uuid)
-                    .withValue(ChartItems.WEIGHT, nextWeight++)
-                    .withValue(ChartItems.PARENT_ROWID, parentId)
-                    .withValue(ChartItems.LABEL, item.label)
-                    .withValue(ChartItems.TYPE, item.type)
-                    .withValue(ChartItems.REQUIRED, item.required ? 1 : 0)
-                    .withValue(ChartItems.CONCEPT_UUIDS, Joiner.on(",").join(conceptUuids))
-                    .withValue(ChartItems.FORMAT, item.format)
-                    .withValue(ChartItems.CAPTION_FORMAT, item.caption_format)
-                    .withValue(ChartItems.CSS_CLASS, item.css_class)
-                    .withValue(ChartItems.CSS_STYLE, item.css_style)
-                    .withValue(ChartItems.SCRIPT, item.script)
-                    .build());
-                syncResult.stats.numInserts++;
-            }
-        }
-        return ops;
-    }
-
-    public static List<ContentProviderOperation> getFormUpdateOps(SyncResult syncResult)
-        throws ExecutionException, InterruptedException {
-        LOG.i("Listing all forms on server");
-        RequestFuture<List<JsonForm>> future = RequestFuture.newFuture();
-        App.getServer().listForms(future, future);
-        Map<String, ContentValues> cvs = new HashMap<>();
-        for (JsonForm form : future.get()) {
-            cvs.put(form.id, Form.fromJson(form).toContentValues());
-        }
-
-        ArrayList<ContentProviderOperation> ops = new ArrayList<>();
-        final ContentResolver resolver = App.getInstance().getContentResolver();
-        Cursor c = resolver.query(Contracts.Forms.CONTENT_URI, null, null, null, null);
-        LOG.i("Examining forms: " + c.getCount() + " local, " + cvs.size() + " from server");
-        try {
-            while (c.moveToNext()) {
-                String uuid = Utils.getString(c, Contracts.Forms.UUID);
-                Uri uri = Contracts.Forms.CONTENT_URI.buildUpon().appendPath(uuid).build();
-                LOG.i("  - will delete form " + uuid);
-                ops.add(ContentProviderOperation.newDelete(uri).build());
-            }
-        } finally {
-            c.close();
-        }
-
-        for (ContentValues values : cvs.values()) {  // server has a new record
-            LOG.i("  - will insert form " + values.getAsString(Contracts.Forms.UUID));
-            ops.add(ContentProviderOperation.newInsert(Contracts.Forms.CONTENT_URI).withValues(values).build());
-            syncResult.stats.numInserts++;
-        }
-        return ops;
-    }
-
-    /**
-     * Downloads all patients from the server and produces a list of the operations
-     * needed to bring the local database in sync with the server.
-     */
-    public static List<ContentProviderOperation> getPatientUpdateOps(SyncResult syncResult)
-        throws ExecutionException, InterruptedException {
-        final ContentResolver resolver = App.getInstance().getContentResolver();
-
-        RequestFuture<List<JsonPatient>> future = RequestFuture.newFuture();
-        App.getServer().listPatients("", "", "", future, future);
-        Map<String, ContentValues> cvs = new HashMap<>();
-        for (JsonPatient patient : future.get()) {
-            cvs.put(patient.id, Patient.fromJson(patient).toContentValues());
-        }
-
-        Cursor c = resolver.query(Patients.CONTENT_URI, null, null, null, null);
-        LOG.i("Examining patients: " + c.getCount() + " local, " + cvs.size() + " from server");
-
-        List<ContentProviderOperation> ops = new ArrayList<>();
-        try {
-            while (c.moveToNext()) {
-                String uuid = Utils.getString(c, Patients.UUID);
-                Uri uri = Patients.CONTENT_URI.buildUpon().appendPath(uuid).build();
-                syncResult.stats.numEntries++;
-
-                ContentValues cv = cvs.remove(uuid);
-                if (cv != null) {
-                    ContentValues localCv = new ContentValues();
-                    DatabaseUtils.cursorRowToContentValues(c, localCv);
-                    if (!cv.equals(localCv)) {  // record has changed on server
-                        LOG.i("  - will update patient " + uuid);
-                        ops.add(ContentProviderOperation.newUpdate(uri).withValues(cv).build());
-                        syncResult.stats.numUpdates++;
-                    }
-                } else {  // record doesn't exist on server
-                    LOG.i("  - will delete patient " + uuid);
-                    ops.add(ContentProviderOperation.newDelete(uri).build());
-                    syncResult.stats.numDeletes++;
-                }
-            }
-        } finally {
-            c.close();
-        }
-
-        for (ContentValues values : cvs.values()) {  // server has a new record
-            LOG.i("  - will insert patient " + values.getAsString(Patients.UUID));
-            ops.add(ContentProviderOperation.newInsert(Patients.CONTENT_URI).withValues(values).build());
-            syncResult.stats.numInserts++;
-        }
-
-        return ops;
-    }
-
-    /** Converts an encounter data response into appropriate inserts in the encounters table. */
-    public static List<ContentValues> getObsValuesToInsert(
-            JsonEncounter encounter, SyncResult syncResult) {
-        List<ContentValues> cvs = new ArrayList<>();
-        ContentValues base = new ContentValues();
-        base.put(Observations.PATIENT_UUID, encounter.patient_uuid);
-        base.put(Observations.ENCOUNTER_UUID, encounter.uuid);
-        base.put(Observations.ENCOUNTER_MILLIS, encounter.timestamp.getMillis());
-
-        if (encounter.observations != null) {
-            for (Map.Entry<Object, Object> entry : encounter.observations.entrySet()) {
-                final String conceptUuid = (String) entry.getKey();
-                ContentValues values = new ContentValues(base);
-                values.put(Observations.CONCEPT_UUID, conceptUuid);
-                values.put(Observations.VALUE, entry.getValue().toString());
-                cvs.add(values);
-                syncResult.stats.numInserts++;
-            }
-        }
-        if (encounter.order_uuids != null) {
-            for (String orderUuid : encounter.order_uuids) {
-                ContentValues values = new ContentValues(base);
-                values.put(Observations.CONCEPT_UUID, AppModel.ORDER_EXECUTED_CONCEPT_UUID);
-                values.put(Observations.VALUE, orderUuid);
-                cvs.add(values);
-                syncResult.stats.numInserts++;
-            }
-        }
-        return cvs;
-    }
-
-    /**
-     * Gets orders from the server and returns a list of operations that will
-     * update the database with the new orders and edits to existing orders.
-     */
-    public static ArrayList<ContentProviderOperation> getOrderUpdateOps(SyncResult syncResult)
-        throws ExecutionException, InterruptedException {
-        // Request all orders from the server.
-        RequestFuture<List<JsonOrder>> future = RequestFuture.newFuture();
-        App.getServer().listOrders(future, future);
-        Map<String, JsonOrder> ordersToStore = new HashMap<>();
-        for (JsonOrder order : future.get()) {
-            ordersToStore.put(order.uuid, order);
-        }
-
-        ArrayList<ContentProviderOperation> ops = new ArrayList<>();
-        final ContentResolver resolver = App.getInstance().getContentResolver();
-        Cursor c = resolver.query(Orders.CONTENT_URI, new String[] {
-            Orders.UUID,
-            Orders.PATIENT_UUID,
-            Orders.INSTRUCTIONS,
-            Orders.START_MILLIS,
-            Orders.STOP_MILLIS
-        }, null, null, null);
-        try {
-            LOG.i("Examining orders: %d local, %d from server.", c.getCount(), ordersToStore.size());
-            // Scan all the locally stored orders, updating the orders we've just received.
-            while (c.moveToNext()) {
-                String uuid = c.getString(c.getColumnIndex(Orders.UUID));
-                Uri uri = Orders.CONTENT_URI.buildUpon().appendPath(uuid).build();
-                JsonOrder order = ordersToStore.get(uuid);
-                if (order != null) {  // apply update to a local order
-                    LOG.v("  - will update order " + uuid);
-                    ops.add(ContentProviderOperation.newUpdate(uri)
-                        .withValue(Orders.PATIENT_UUID, order.patient_uuid)
-                        .withValue(Orders.INSTRUCTIONS, order.instructions)
-                        .withValue(Orders.START_MILLIS, order.start_millis)
-                        .withValue(Orders.STOP_MILLIS, order.stop_millis)
-                        .build());
-                    ordersToStore.remove(uuid);  // done with this incoming order
-                    syncResult.stats.numUpdates++;
-                } else {  // delete the local order (the server doesn't have it)
-                    LOG.v("  - will delete order " + uuid);
-                    ops.add(ContentProviderOperation.newDelete(uri).build());
-                    syncResult.stats.numDeletes++;
-                }
-            }
-        } finally {
-            c.close();
-        }
-
-        // Store all the remaining received orders as new orders.
-        for (JsonOrder order : ordersToStore.values()) {
-            LOG.v("  - will insert order " + order.uuid);
-            ops.add(ContentProviderOperation.newInsert(Orders.CONTENT_URI)
-                .withValue(Orders.UUID, order.uuid)
-                .withValue(Orders.PATIENT_UUID, order.patient_uuid)
-                .withValue(Orders.INSTRUCTIONS, order.instructions)
-                .withValue(Orders.START_MILLIS, order.start_millis)
-                .withValue(Orders.STOP_MILLIS, order.stop_millis)
-                .build());
-            syncResult.stats.numInserts++;
-        }
-        return ops;
-    }
-
     /** Given a set of users, replaces the current set of users with users from that set. */
     public static ArrayList<ContentProviderOperation> getUserUpdateOps(
         Set<JsonUser> response, SyncResult syncResult) {
@@ -305,153 +78,5 @@ public class DbSyncHelper {
             syncResult.stats.numInserts++;
         }
         return ops;
-    }
-
-    /**
-     * Requests locations from the server and transforms the response into an {@link ArrayList} of
-     * {@link ContentProviderOperation}s for updating the database.
-     */
-    public static ArrayList<ContentProviderOperation> getLocationUpdateOps(SyncResult syncResult)
-        throws ExecutionException, InterruptedException {
-        final ContentResolver contentResolver = App.getInstance().getContentResolver();
-
-        final String[] projection = new String[] {
-            Locations.UUID,
-            Locations.PARENT_UUID
-        };
-        final String[] namesProjection = new String[] {
-            LocationNames.LOCATION_UUID,
-            LocationNames.LOCALE,
-            LocationNames.NAME
-        };
-
-        LOG.d("Before network call");
-        RequestFuture<List<JsonLocation>> future = RequestFuture.newFuture();
-        App.getServer().listLocations(future, future);
-
-        // No need for callbacks as the {@AbstractThreadedSyncAdapter} code is executed in a
-        // background thread
-        List<JsonLocation> locations = future.get();
-        LOG.d("After network call");
-        ArrayList<ContentProviderOperation> batch = new ArrayList<>();
-
-        Map<String, JsonLocation> locationsByUuid = new HashMap<>();
-        for (JsonLocation location : locations) {
-            locationsByUuid.put(location.uuid, location);
-        }
-
-        // Get list of all items
-        Uri uri = Locations.CONTENT_URI; // Location tree
-        Uri namesUri = LocationNames.CONTENT_URI; // Location names
-        Cursor c = contentResolver.query(uri, projection, null, null, null);
-        assert c != null;
-        Cursor namesCur = contentResolver.query(namesUri, namesProjection, null, null, null);
-        assert namesCur != null;
-        LOG.i("Examining locations: %d local, %d from server", c.getCount(), locations.size());
-
-        String uuid;
-        String parentUuid;
-
-        // Build map of location names from the database.
-        Map<String, Map<String, String>> dbLocationNames = new HashMap<>();
-        while (namesCur.moveToNext()) {
-            String locationUuid = namesCur.getString(
-                namesCur.getColumnIndex(LocationNames.LOCATION_UUID));
-            String locale = namesCur.getString(
-                namesCur.getColumnIndex(LocationNames.LOCALE));
-            String name = namesCur.getString(
-                namesCur.getColumnIndex(LocationNames.NAME));
-            if (locationUuid == null || locale == null || name == null) continue;
-
-            if (!dbLocationNames.containsKey(locationUuid)) {
-                dbLocationNames.put(locationUuid, new HashMap<String, String>());
-            }
-
-            dbLocationNames.get(locationUuid).put(locale, name);
-        }
-        namesCur.close();
-
-        // Iterate through the list of locations
-        while (c.moveToNext()) {
-            syncResult.stats.numEntries++;
-
-            uuid = c.getString(c.getColumnIndex(Locations.UUID));
-            parentUuid = c.getString(c.getColumnIndex(Locations.PARENT_UUID));
-
-            JsonLocation location = locationsByUuid.get(uuid);
-            if (location != null) {
-                // Entry exists. Remove from entry map to prevent insert later.
-                locationsByUuid.remove(uuid);
-
-                // Grab the names stored in the database for this location.
-                Map<String, String> locationNames = dbLocationNames.get(uuid);
-
-                // Check to see if the entry needs to be updated
-                Uri existingUri = uri.buildUpon().appendPath(String.valueOf(uuid)).build();
-
-                if (location.parent_uuid != null && !location.parent_uuid.equals(parentUuid)) {
-                    // Update existing record
-                    LOG.i("  - will update location " + uuid);
-                    batch.add(ContentProviderOperation.newUpdate(existingUri)
-                        .withValue(Locations.UUID, uuid)
-                        .withValue(Locations.PARENT_UUID, parentUuid)
-                        .build());
-                    syncResult.stats.numUpdates++;
-                }
-
-                if (location.names != null
-                    && (locationNames == null || !location.names.equals(locationNames))) {
-                    Uri existingNamesUri = namesUri.buildUpon().appendPath(
-                        String.valueOf(uuid)).build();
-                    // Update location names by deleting any existing location names and
-                    // repopulating.
-                    batch.add(ContentProviderOperation.newDelete(existingNamesUri).build());
-                    syncResult.stats.numDeletes++;
-                    for (String locale : location.names.keySet()) {
-
-                        batch.add(ContentProviderOperation.newInsert(existingNamesUri)
-                            .withValue(LocationNames.LOCATION_UUID, uuid)
-                            .withValue(LocationNames.LOCALE, locale)
-                            .withValue(LocationNames.NAME, location.names.get(locale))
-                            .build());
-                        syncResult.stats.numInserts++;
-                    }
-                }
-            } else {
-                // Entry doesn't exist. Remove it from the database.
-                LOG.i("  - will delete location " + uuid);
-                Uri deleteUri = uri.buildUpon().appendPath(uuid).build();
-                batch.add(ContentProviderOperation.newDelete(deleteUri).build());
-                syncResult.stats.numDeletes++;
-                Uri namesDeleteUri = namesUri.buildUpon().appendPath(uuid).build();
-                batch.add(ContentProviderOperation.newDelete(namesDeleteUri).build());
-                syncResult.stats.numDeletes++;
-            }
-        }
-        c.close();
-
-        for (JsonLocation location : locationsByUuid.values()) {
-            LOG.i("  - will insert location " + location.uuid);
-            batch.add(ContentProviderOperation.newInsert(Locations.CONTENT_URI)
-                .withValue(Locations.UUID, location.uuid)
-                .withValue(Locations.PARENT_UUID, location.parent_uuid)
-                .build());
-            syncResult.stats.numInserts++;
-
-            if (location.names != null) {
-                for (String locale : location.names.keySet()) {
-                    Uri existingNamesUri = namesUri.buildUpon().appendPath(
-                        String.valueOf(location.uuid)).build();
-                    batch.add(ContentProviderOperation.newInsert(existingNamesUri)
-                        .withValue(LocationNames.LOCATION_UUID, location.uuid)
-                        .withValue(LocationNames.LOCALE, locale)
-                        .withValue(LocationNames.NAME, location.names.get(locale))
-                        .build());
-                    syncResult.stats.numInserts++;
-                }
-            }
-        }
-
-        return batch;
     }
 }
